+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(MANOVA.RM)
+
+test_check("MANOVA.RM")